Converted data types to more explicit naming-convention.

+++ Avail.h
@@ -7,28 +7,28 @@ class Avail {
 
 	public:
 
-		static bool micros(unsigned long *margin, unsigned long *lastMicros, bool &rollover) {
+		static bool micros(uint32_t *margin, uint32_t *lastMicros, bool &rollover) {
 			return Avail::_avail(::micros(), margin, lastMicros, rollover);
 		};
 
-		static bool micros(unsigned long *margin, unsigned long *lastMicros) {
+		static bool micros(uint32_t *margin, uint32_t *lastMicros) {
 			bool _;
 			return Avail::micros(margin, lastMicros, _);
 		};
 
-		static bool millis(unsigned long *margin, unsigned long *lastMillis, bool &rollover) {
+		static bool millis(uint32_t *margin, uint32_t *lastMillis, bool &rollover) {
 			return Avail::_avail(::millis(), margin, lastMillis, rollover);
 		};
 
-		static bool millis(unsigned long *margin, unsigned long *lastMillis) {
+		static bool millis(uint32_t *margin, uint32_t *lastMillis) {
 			bool _;
 			return Avail::millis(margin, lastMillis, _);
 		};
 
 	private:
 	
-		static bool _avail(unsigned long metric, unsigned long *margin, unsigned long *lastLong, bool &rollover) {
-			unsigned long diff = metric - *lastLong;
+		static bool _avail(uint32_t metric, uint32_t *margin, uint32_t *lastLong, bool &rollover) {
+			uint32_t diff = metric - *lastLong;
 			rollover = diff > metric;
 			if (rollover) { diff = 4294967295 - *lastLong + metric; }
